requirements.txt
art==4.5.*
PyGithub==1.*